Use new index to iterate through assets backed by a certain asset.
