+++ tests/tests/bitasset_tests.cpp
@@ -80,7 +80,7 @@ void change_backing_asset(database_fixture& fixture, const fc::ecc::private_key&
 }
 
 /******
- * @ brief helper method to turn witness_fed_asset on and off
+ * @brief helper method to turn witness_fed_asset on and off
  * @param fixture the database_fixture
  * @param new_issuer optionally change the issuer
  * @param signing_key signer
@@ -114,6 +114,26 @@ void change_asset_options(database_fixture& fixture, const optional<account_id_t
    fixture.trx.clear();
 
 }
+
+/*********
+ * @brief helper method to create a coin backed by a bitasset
+ * @param fixture the database_fixture
+ * @param index added to name of the coin
+ * @param backing the backing asset
+ * @param signing_key the signing key
+ */
+const graphene::chain::asset_object& create_bitasset_backed(graphene::chain::database_fixture& fixture,
+      int index, graphene::chain::asset_id_type backing, const fc::ecc::private_key& signing_key)
+{
+   // create the coin
+   std::string name = "COIN" + std::to_string(index + 1) + "TEST";
+   const graphene::chain::asset_object& obj = fixture.create_bitasset(name);
+   // adjust the backing asset
+   change_backing_asset(fixture, signing_key, obj.get_id(), backing);
+   fixture.trx.set_expiration(fixture.db.get_dynamic_global_properties().next_maintenance_time);
+   return obj;
+}
+
 
 /*********
  * @brief make sure feeds still work after changing backing asset on a witness-fed asset
@@ -1108,6 +1128,78 @@ BOOST_AUTO_TEST_CASE( hf_935_test )
       }
    }
 } FC_LOG_AND_RETHROW() }
+
+BOOST_AUTO_TEST_CASE( bitasset_secondary_index )
+{
+   ACTORS( (nathan) );
+
+   graphene::chain::asset_id_type core_id;
+   BOOST_TEST_MESSAGE("Create coins");
+   try
+   {
+      // make 5 coins (backed by core)
+      for(int i = 0; i < 5; i++)
+      {
+         create_bitasset_backed(*this, i, core_id, nathan_private_key);
+      }
+      // make the next 5 (10-14) be backed by COIN1
+      graphene::chain::asset_id_type coin1_id = get_asset("COIN1TEST").get_id();
+      for(int i = 5; i < 10; i++)
+      {
+         create_bitasset_backed(*this, i, coin1_id, nathan_private_key);
+      }
+      // make the next 5 (15-19) be backed by COIN2
+      graphene::chain::asset_id_type coin2_id = get_asset("COIN2TEST").get_id();
+      for(int i = 10; i < 15; i++)
+      {
+         create_bitasset_backed(*this, i, coin2_id, nathan_private_key);
+      }
+      // make the last 5 be backed by core
+      for(int i = 15; i < 20; i++)
+      {
+         create_bitasset_backed(*this, i, core_id, nathan_private_key);
+      }
+
+      BOOST_TEST_MESSAGE("Searching for all coins backed by CORE");
+      const auto& idx = db.get_index_type<graphene::chain::asset_bitasset_data_index>().indices().get<graphene::chain::by_short_backing_asset>();
+      auto core_itr = idx.find( core_id );
+      auto core_end = idx.upper_bound(core_id);
+      BOOST_TEST_MESSAGE("Searching for all coins backed by COIN1");
+      auto coin1_itr = idx.find( coin1_id );
+      auto coin1_end = idx.upper_bound( coin1_id );
+      BOOST_TEST_MESSAGE("Searching for all coins backed by COIN2");
+      auto coin2_itr = idx.find( coin2_id );
+
+      int core_count = 0, coin1_count = 0, coin2_count = 0;
+
+      BOOST_TEST_MESSAGE("Counting coins in each category");
+
+      for( ; core_itr != core_end; ++core_itr)
+      {
+         BOOST_CHECK(core_itr->options.short_backing_asset == core_id);
+         core_count++;
+      }
+      for( ; coin1_itr != coin1_end; ++coin1_itr )
+      {
+         BOOST_CHECK(coin1_itr->options.short_backing_asset == coin1_id);
+         coin1_count++;
+      }
+      for( ; coin2_itr != idx.end(); ++coin2_itr )
+      {
+         BOOST_CHECK(coin2_itr->options.short_backing_asset == coin2_id);
+         coin2_count++;
+      }
+
+      BOOST_CHECK_EQUAL(core_count, 10);
+      BOOST_CHECK_EQUAL(coin1_count, 5);
+      BOOST_CHECK_EQUAL(coin2_count, 5);
+   }
+   catch (fc::exception& ex)
+   {
+      BOOST_FAIL(ex.to_string(fc::log_level(fc::log_level::all)));
+   }
+}
+
 
 /*****
  * @brief make sure feeds work correctly after changing from non-witness-fed to witness-fed before the 868 fork
